Move schema file to reponseSchema
