+++ controllers/items/list.js
@@ -4,7 +4,7 @@
 const { forEach } = require('p-iteration');
 //const Boom = require('boom');
 // const server = require('../../server');
-const Schema = require('../../lib/schema');
+const Schema = require('../../lib/responseSchema');
 
 const swagger = Schema.generate(['404']);
