+++ controllers/users/destroy.js
@@ -3,7 +3,7 @@
 const Joi = require('joi');
 const Boom = require('boom');
 // const Server = require('../../server');
-const Schema = require('../../lib/schema');
+const Schema = require('../../lib/responseSchema');
 
 const swagger = Schema.generate(['404', '401']);
