+++ controllers/users/get_lists.js
@@ -4,7 +4,7 @@ const Joi = require('joi');
 const Boom = require('boom');
 const uuid = require('uuid').v4;
 
-const Schema = require('../../lib/schema');
+const Schema = require('../../lib/responseSchema');
 
 const swagger = Schema.generate();
