+++ controllers/users/get_profile.js
@@ -2,7 +2,7 @@
 
 const Joi = require('joi');
 //const Boom = require('boom');
-const Schema = require('../../lib/schema');
+const Schema = require('../../lib/responseSchema');
 
 const swagger = Schema.generate(['404']);
