+++ controllers/users/update.js
@@ -4,7 +4,7 @@
 //const Config = require('getconfig');
 const Joi = require('joi');
 const Boom = require('boom');
-const Schema = require('../../lib/schema');
+const Schema = require('../../lib/responseSchema');
 
 const swagger = Schema.generate(['404', '409']);
